backend/requirements.txt
numpy==2.2.6
pandas==2.3.3
scipy==1.14.1
matplotlib==3.10.7
seaborn==0.13.2
scikit-learn==1.7.2
tensorflow==2.20.0
torch==2.9.1

opencv-python
flask
flask-cors
pillow
